+++ packages/gatsby-source-contentful/src/__tests__/fetch.js
@@ -84,6 +84,7 @@ beforeAll(() => {
 })
 
 const reporter = {
+  info: jest.fn(),
   panic: jest.fn(),
 }