+++ packages/gatsby/src/redux/reducers/index.js
@@ -1,9 +1,18 @@
 const reduxNodes = require(`./nodes`)
 const lokiNodes = require(`../../db/loki/nodes`).reducer
+import { pagesReducer } from "./pages"
 import { redirectsReducer } from "./redirects"
+import { schemaReducer } from "./schema"
+import { staticQueryComponentsReducer } from "./static-query-components"
+import { statusReducer } from "./status"
+import { webpackReducer } from "./webpack"
+import { pageDataReducer } from "./page-data"
+import { themesReducer } from "./themes"
+import { webpackCompilationHashReducer } from "./webpack-compilation-hash"
 import { reducer as logReducer } from "gatsby-cli/lib/reporter/redux/reducer"
 
-const backend = process.env.GATSBY_DB_NODES || `redux`
+// const backend = process.env.GATSBY_DB_NODES || `redux`
+const backend = `redux`
 
 function getNodesReducer() {
   let nodesReducer
@@ -51,22 +60,22 @@ module.exports = {
   lastAction: require(`./last-action`),
   flattenedPlugins: require(`./flattened-plugins`),
   config: require(`./config`),
-  pages: require(`./pages`),
-  schema: require(`./schema`),
-  status: require(`./status`),
+  schema: schemaReducer,
+  pages: pagesReducer,
+  status: statusReducer,
   componentDataDependencies: require(`./component-data-dependencies`),
   components: require(`./components`),
-  staticQueryComponents: require(`./static-query-components`),
+  staticQueryComponents: staticQueryComponentsReducer,
   jobs: require(`./jobs`),
   jobsV2: require(`./jobsv2`),
-  webpack: require(`./webpack`),
-  webpackCompilationHash: require(`./webpack-compilation-hash`),
+  webpack: webpackReducer,
+  webpackCompilationHash: webpackCompilationHashReducer,
   redirects: redirectsReducer,
   babelrc: require(`./babelrc`),
   schemaCustomization: require(`./schema-customization`),
-  themes: require(`./themes`),
+  themes: themesReducer,
   logs: logReducer,
   inferenceMetadata: require(`./inference-metadata`),
   pageDataStats: require(`./page-data-stats`),
-  pageData: require(`./page-data`),
+  pageData: pageDataReducer,
 }