+++ www/src/components/homepage/homepage-features.js
@@ -50,9 +50,8 @@ const HomepageFeatures = () => (
     <Card>
       <CardHeadline>Future-proof your website</CardHeadline>
       <FuturaParagraph>
-        Do not build a website with last decade’s tech. The future of the web is
-        mobile, JavaScript and APIs—the {` `}
-        <a href="https://jamstack.org/">JAMstack</a>. Every website is a web app
+        Don't build a website with last decade’s tech. The future of the web is
+        mobile, JavaScript and APIs. Every website is a web app
         and every web app is a website. Gatsby.js is the universal JavaScript
         framework you’ve been waiting for.
       </FuturaParagraph>